Rename overlay to modal
