+++ src/ignitus-Routes/ignitus-UserInterfaceBookRoutes/index.tsx
@@ -22,7 +22,7 @@ import { interfaceSeparator } from '../../ignitus-UserInterfaceBook/Components/A
 import { interfaceDashBoardFooter } from '../../ignitus-UserInterfaceBook/Components/Organisms/interfaceFooter/interfaceDashBoardFooter/Components';
 import { interfaceIconInput } from '../../ignitus-UserInterfaceBook/Components/Atoms/interfaceIconInput/Components';
 import { interfaceMultiMediaInput } from '../../ignitus-UserInterfaceBook/Components/Atoms/interfaceMulti/Components';
-import { interfaceOverlay } from '../../ignitus-UserInterfaceBook/Components/Molecules/interfaceOverlay/Components';
+import { interfaceModal } from '../../ignitus-UserInterfaceBook/Components/Molecules/interfaceModal';
 import { interfaceTags } from '../../ignitus-UserInterfaceBook/Components/Atoms/interfaceTags/Components';
 import { interfaceFilters } from '../../ignitus-UserInterfaceBook/Components/Templates/interfaceFilters/Components/index';
 import { interfaceUserProfile } from '../../ignitus-UserInterfaceBook/Components/Organisms/interfaceUserProfile/Components';
@@ -128,7 +128,7 @@ export const UserInterfaceBookRoutes: React.FunctionComponent = () => (
         />
         <Route path="/interface/filters" component={interfaceFilters} />
         <Route path="/interface/userProfile" component={interfaceUserProfile} />
-        <Route path="/interface/overlay" component={interfaceOverlay} />
+        <Route path="/interface/modal" component={interfaceModal} />
         <Route path="/interface/scrollBar" component={interfaceScrollBar} />
         <Route
           path="/interface/toggleButtons"
