+++ src/ignitus-Shared/index.ts
@@ -67,11 +67,11 @@ export {
 /**
  * ignitus-Molecules
  */
-// Because Overlay imports Heading1 from this one, therefore disabling no-cycle wont create any harm
+// Because Modal imports Heading1 from this one, therefore disabling no-cycle wont create any harm
 // eslint-disable-next-line import/no-cycle
 export {
-  Overlay,
-} from './ignitus-DesignSystem/ignitus-Molecules/ignitus-Overlay/Components';
+  Modal,
+} from './ignitus-DesignSystem/ignitus-Molecules/ignitus-Modal/Components';
 
 export {
   ScrollBar,
