+++ src/ignitus-UserInterfaceBook/Components/Molecules/interfaceModal/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { Overlay, Heading2 } from '../../../../../ignitus-Shared';
-import { Interface } from '../../../../styles';
+import { Modal, Heading2 } from '../../../../ignitus-Shared';
+import { Interface } from '../../../styles';
 
-export const interfaceOverlay = () => (
+export const interfaceModal = () => (
   <Interface>
-    <Heading2>Overlay</Heading2>
+    <Heading2>Modal</Heading2>
     <hr />
-    <Overlay
+    <Modal
       primaryText="Your application has been sent successfully!"
       secondaryText="Make sure to keep an eye on your email."
       buttonText="Go Back"
